[NCL-5611] Initial commit to port to protobuf
This commit adds classes to properly serialize class:

- MultiValuedMap
- LocalFile

to Protobuf format. `ProtobufSerializer` is the interface annotated with
classes to be considered by the Protobuf Stream processor (Protostream)
to generate proto files for serialization and deserialization.

The Protobuf format is not yet 'activated' in the codebase as we work on
the rest of the classes that need to be serialized via Protobuf.

The spotbug-exclude.xml includes the classes generated by Protostream so
that they are not scanned for checks