[NCL-6344] Implement Supplier interface
Implement Supplier interface for `BuildFinder` and
`DistributionAnalyzer`. This is useful when invoking both of them in a
`CompletableFuture.supplyAsync` to produce a `CompletableFuture`. The
latter can then be composed and combined with other `CompletableFuture`s.

One reason for using that composability for NCL-6344 is to implement a
cancellation feature. The flow goes from:

url -> DistributionAnalyzer -> BuildFinder

When running `DistributionAnalyzer` and `BuildFinder` as a future, it is
really impossible to stop them in case of cancellation. However, we can
cancel the rest of the flow after them.

url -> future(DistributionAnalyzer) -> check if cancelled -> future(BuildFinder)

Using a `CompletableFuture` makes this much easier.