[NCL-6344] Add explicit cancel checkpoints
This commit adds cancel support for BuildFinder and
DistributionAnalyzer. They are both Callables and are typically run
inside a threadpool.

While we could use `future.cancel(true)` to force the callable to stop,
it just sends the interrupt signal to the thread running the callable
and it's up to the callable to check whether the thread is interrupted,
and stop.

This commit attempts to do something similar to the interrupt signal
with the `CancelWrapper`. It is a simple wrapper around an AtomicBoolean
that is set to true if we want to cancel/stop everything. It is passed
to the callable via:

```java
// user specifies the cancel wrapper object
buildFinder.setCancelWrapper(cancelWrapper);

...

buildFinder.call();
```

At various checkpoints in the 2 callables, we do an explicit check to
see if the cancelWrapper is true. If true, we stop and throw a
`CancelException`, which is just a `RuntimeException`.

We prefer to use the `CancelWrapper` instead of the Interrupt signal
because the latter is not sent when we try to cancel a
`CompletableFuture`.